[refactor] separate Express server and WebSocket server into different files
